Fix Battle Field and Second Level Activity
