+++ ColourMemory/app/src/main/AndroidManifest.xml
@@ -8,7 +8,6 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
         <activity
-            android:screenOrientation="portrait"
             android:name=".MainActivity"
             android:label="@string/app_name">
             <intent-filter>
@@ -17,6 +16,9 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
+        <activity android:name=".ui.HighScoreActivity"
+            android:parentActivityName=".MainActivity"/>
     </application>
 
 </manifest>
