+++ ColourMemory/app/src/main/res/layout/activity_high_score.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+              android:layout_width="match_parent"
+              android:layout_height="match_parent">
+
+    <TextView
+        android:id="@+id/tv_highscore_title"
+        android:gravity="center_horizontal"
+        style="@style/Arial.YellowOrange.H1"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="@string/title_high_scores"/>
+
+    <ListView
+        android:layout_below="@+id/tv_highscore_title"
+        android:id="@+id/lv_scores"
+        android:layout_width="wrap_content"
+        android:layout_height="match_parent"></ListView>
+
+
+
+</RelativeLayout>
