+++ ColourMemory/app/src/main/res/layout/fragment_card_board.xml
@@ -2,13 +2,14 @@
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                android:orientation="vertical">
+                android:orientation="vertical"
+                android:padding="8dp">
 
     <FrameLayout
         android:id="@+id/fl_header"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginBottom="20dp">
+        android:layout_marginBottom="10dp">
 
         <include layout="@layout/view_header"/>
     </FrameLayout>
@@ -16,23 +17,147 @@
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_below="@+id/fl_header">
-
-        <LinearLayout
-            android:id="@+id/container"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:orientation="vertical">
-
-        </LinearLayout>
-    </FrameLayout>
-
-    <FrameLayout
         android:layout_below="@+id/fl_header"
-        android:id="@+id/fl_blocker"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:visibility="invisible"
-        android:background="@android:color/transparent"/>
+        android:layout_centerInParent="true">
+
+        <TableLayout
+
+            android:id="@+id/tbl_field"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content">
+
+            <TableRow
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content">
+
+                <com.plex.jlrf.colourmemory.model.Card
+                    android:layout_width="@dimen/card_width"
+                    android:layout_height="@dimen/card_height"
+                    android:layout_margin="@dimen/card_margin"
+                    android:layout_weight="1"/>
+
+                <com.plex.jlrf.colourmemory.model.Card
+                    android:id="@+id/card_2"
+                    android:layout_width="@dimen/card_width"
+                    android:layout_height="@dimen/card_height"
+                    android:layout_margin="@dimen/card_margin"
+                    android:layout_weight="1"/>
+
+                <com.plex.jlrf.colourmemory.model.Card
+                    android:id="@+id/card_3"
+                    android:layout_width="@dimen/card_width"
+                    android:layout_height="@dimen/card_height"
+                    android:layout_margin="@dimen/card_margin"
+                    android:layout_weight="1"/>
+
+                <com.plex.jlrf.colourmemory.model.Card
+                    android:id="@+id/card_4"
+                    android:layout_width="@dimen/card_width"
+                    android:layout_height="@dimen/card_height"
+                    android:layout_margin="@dimen/card_margin"
+                    android:layout_weight="1"/>
+            </TableRow>
+
+            <TableRow
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content">
+
+                <com.plex.jlrf.colourmemory.model.Card
+                    android:id="@+id/card_5"
+                    android:layout_width="@dimen/card_width"
+                    android:layout_height="@dimen/card_height"
+                    android:layout_margin="@dimen/card_margin"
+                    android:layout_weight="1"/>
+
+                <com.plex.jlrf.colourmemory.model.Card
+                    android:id="@+id/card_6"
+                    android:layout_width="@dimen/card_width"
+                    android:layout_height="@dimen/card_height"
+                    android:layout_margin="@dimen/card_margin"
+                    android:layout_weight="1"/>
+
+                <com.plex.jlrf.colourmemory.model.Card
+                    android:id="@+id/card_7"
+                    android:layout_width="@dimen/card_width"
+                    android:layout_height="@dimen/card_height"
+                    android:layout_margin="@dimen/card_margin"
+                    android:layout_weight="1"/>
+
+                <com.plex.jlrf.colourmemory.model.Card
+                    android:id="@+id/card_8"
+                    android:layout_width="@dimen/card_width"
+                    android:layout_height="@dimen/card_height"
+                    android:layout_margin="@dimen/card_margin"
+                    android:layout_weight="1"/>
+            </TableRow>
+
+            <TableRow
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content">
+
+                <com.plex.jlrf.colourmemory.model.Card
+                    android:id="@+id/card_9"
+                    android:layout_width="@dimen/card_width"
+                    android:layout_height="@dimen/card_height"
+                    android:layout_margin="@dimen/card_margin"
+                    android:layout_weight="1"/>
+
+                <com.plex.jlrf.colourmemory.model.Card
+                    android:id="@+id/card_10"
+                    android:layout_width="@dimen/card_width"
+                    android:layout_height="@dimen/card_height"
+                    android:layout_margin="@dimen/card_margin"
+                    android:layout_weight="1"/>
+
+                <com.plex.jlrf.colourmemory.model.Card
+                    android:id="@+id/card_11"
+                    android:layout_width="@dimen/card_width"
+                    android:layout_height="@dimen/card_height"
+                    android:layout_margin="@dimen/card_margin"
+                    android:layout_weight="1"/>
+
+                <com.plex.jlrf.colourmemory.model.Card
+                    android:id="@+id/card_12"
+                    android:layout_width="@dimen/card_width"
+                    android:layout_height="@dimen/card_height"
+                    android:layout_margin="@dimen/card_margin"
+                    android:layout_weight="1"/>
+            </TableRow>
+
+            <TableRow
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content">
+
+                <com.plex.jlrf.colourmemory.model.Card
+                    android:id="@+id/card_13"
+                    android:layout_width="@dimen/card_width"
+                    android:layout_height="@dimen/card_height"
+                    android:layout_margin="@dimen/card_margin"
+                    android:layout_weight="1"/>
+
+                <com.plex.jlrf.colourmemory.model.Card
+                    android:id="@+id/card_14"
+                    android:layout_width="@dimen/card_width"
+                    android:layout_height="@dimen/card_height"
+                    android:layout_margin="@dimen/card_margin"
+                    android:layout_weight="1"/>
+
+                <com.plex.jlrf.colourmemory.model.Card
+                    android:id="@+id/card_15"
+                    android:layout_width="@dimen/card_width"
+                    android:layout_height="@dimen/card_height"
+                    android:layout_margin="@dimen/card_margin"
+                    android:layout_weight="1"/>
+
+                <com.plex.jlrf.colourmemory.model.Card
+                    android:id="@+id/card_16"
+                    android:layout_width="@dimen/card_width"
+                    android:layout_height="@dimen/card_height"
+                    android:layout_margin="@dimen/card_margin"
+                    android:layout_weight="1"/>
+            </TableRow>
+        </TableLayout>
+
+    </FrameLayout>
 
 </RelativeLayout>
