--- ColourMemory/app/src/main/res/layout/fragment_high_score.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              android:layout_width="match_parent"
-              android:layout_height="match_parent">
-
-    <FrameLayout
-        android:id="@+id/fl_header"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-        <TextView
-            android:gravity="center_horizontal"
-            style="@style/Arial.YellowOrange.H1"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="@string/title_high_scores"/>
-
-    </FrameLayout>
-
-    <FrameLayout
-        android:layout_below="@+id/fl_header"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-        <ListView
-            android:id="@+id/lv_scores"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"></ListView>
-
-    </FrameLayout>
-
-
-</RelativeLayout>
