+++ ColourMemory/app/src/main/res/layout/item_score.xml
@@ -5,18 +5,20 @@
               android:layout_width="match_parent"
               android:layout_height="match_parent">
     <TextView
-        style="@style/Arial.White"
+        style="@style/Arial.Yellow"
         android:id="@+id/tv_nickname"
         tools:text="Jose Luis Ramos Fernandez"
         android:layout_weight="1"
+        android:gravity="left"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"/>
     <TextView
-        style="@style/Arial.White"
+        style="@style/Arial.Red"
         android:id="@+id/tv_score_value"
         tools:text="11"
+        android:layout_gravity="right"
         android:gravity="right"
-        android:layout_weight="1"
+        android:layout_weight="2"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"/>
 </LinearLayout>
