--- ColourMemory/app/src/main/res/layout/view_card_template.xml
@@ -1,8 +1,0 @@
-<com.plex.jlrf.colourmemory.model.Card
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/card"
-    android:layout_width="60dp"
-    android:layout_height="70dp"
-    android:layout_margin="8dp"
-
-    />
